# apps/docs/public/openapi.yaml
openapi: 3.1.0
info:
  title: Autlantic Billing API
  version: "2026-01-01"
  summary: USDC subscriptions, one-time payments, and payment links on Base
  description: |
    Hosted Billing API for Autlantic Payments.

    - **Test** keys (`abk_test_…`) → Base Sepolia
    - **Live** keys (`abk_live_…`) → Base mainnet
    - **API version** pin: send `Autlantic-Version: 2026-01-01` (default if omitted)

    Production base URL: `https://billing.autlantic.com`

    Authenticate with `x-autlantic-api-key` (or `Authorization: Bearer`).
    Pass `Idempotency-Key` on POST requests to safely retry creates.

    Mobile apps must not use secret API keys. Create sessions on your server and
    open the returned `checkoutUrl` (or payment link `url`) in the app.
  contact:
    name: Autlantic Limited
    url: https://autlantic.com/about
    email: support@autlantic.com
  license:
    name: MIT
servers:
  - url: https://billing.autlantic.com
    description: Production
  - url: http://localhost:8788
    description: Local billing-api
...
security:
  - ApiKeyAuth: []
paths:
...
  /v1/subscriptions:
    get:
      tags: [Subscriptions]
      summary: List subscriptions
      operationId: listSubscriptions
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [incomplete, active, past_due, canceled, paused]
      responses:
        "200":
          description: Subscriptions for this merchant/mode
          content:
            application/json:
              schema:
                type: object
                properties:
                  subscriptions:
                    type: array
                    items:
                      $ref: "#/components/schemas/Subscription"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Subscriptions]
      summary: Create a subscription
      operationId: createSubscription
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubscriptionRequest"
            examples:
              withPriceId:
                value:
                  merchantRef: order_123
                  customerWallet: "0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb0"
                  priceId: price_xxx
              adHoc:
                value:
                  merchantRef: order_123
                  customerWallet: "0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb0"
                  payoutAddressEvm: "0x1111111111111111111111111111111111111111"
                  amountUsdc: 20
                  interval: month
      responses:
        "201":
          description: Incomplete subscription + open invoice + checkout URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  subscription:
                    $ref: "#/components/schemas/Subscription"
                  invoice:
                    $ref: "#/components/schemas/Invoice"
                  checkoutUrl:
                    type: string
                    format: uri
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /v1/payments:
    post:
      tags: [Payments]
      summary: Create a one-time payment
      operationId: createPayment
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentRequest"
      responses:
        "201":
          description: Open payment + checkout URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  payment:
                    $ref: "#/components/schemas/Payment"
                  checkoutUrl:
                    type: string
                    format: uri
...
  /v1/payment-links:
    get:
      tags: [Payment links]
      summary: List payment links
      operationId: listPaymentLinks
      responses:
        "200":
          description: Payment links
          content:
            application/json:
              schema:
                type: object
                properties:
                  paymentLinks:
                    type: array
                    items:
                      $ref: "#/components/schemas/PaymentLink"
    post:
      tags: [Payment links]
      summary: Create a payment link
      operationId: createPaymentLink
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentLinkRequest"
      responses:
        "201":
          description: Link + shareable URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  paymentLink:
                    $ref: "#/components/schemas/PaymentLink"
                  url:
                    type: string
                    format: uri
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-autlantic-api-key
      description: Portal API key (`abk_test_…` or `abk_live_…`)
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
      description: Optional. Cached for 24 hours on POST creates.
...
  responses:
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BadRequest:
      description: Validation or business rule error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        requestId:
          type: string
          description: Server or client request id when available
      required: [error]
...
    Subscription:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
        amountUsdc:
          type: number
        interval:
          type: string
        walletAddress:
          type: string
        currentPeriodEnd:
          type: string
          format: date-time
        cancelAtPeriodEnd:
          type: boolean
    Invoice:
      type: object
      properties:
        id:
          type: string
        subscriptionId:
          type: string
        status:
          type: string
        amountUsdc:
          type: number
        failureCode:
          type: string
          nullable: true
        failureMessage:
          type: string
          nullable: true
        paidAt:
          type: string
          format: date-time
          nullable: true
    Payment:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
        amountUsdc:
          type: number
        customerWallet:
          type: string
        merchantRef:
          type: string
        checkoutUrl:
          type: string
    PaymentLink:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
        amountUsdc:
          type: number
        description:
          type: string
          nullable: true
        useCount:
          type: integer
        maxUses:
          type: integer
          nullable: true
    CreateSubscriptionRequest:
      type: object
      required: [merchantRef, customerWallet]
      properties:
        merchantRef:
          type: string
        customerWallet:
          type: string
        payoutAddressEvm:
          type: string
        amountUsdc:
          type: number
        interval:
          type: string
          enum: [month, year]
        priceId:
          type: string
        planId:
          type: string
        successUrl:
          type: string
          description: Return URL after successful checkout (app deep link or https)
        cancelUrl:
          type: string
          description: Return URL if the customer cancels
        metadata:
          type: object
          additionalProperties:
            type: string
    CreatePaymentRequest:
      type: object
      required: [merchantRef, customerWallet]
      properties:
        merchantRef:
          type: string
        customerWallet:
          type: string
        payoutAddressEvm:
          type: string
        amountUsdc:
          type: number
        priceId:
          type: string
        successUrl:
          type: string
          description: Return URL after successful checkout (app deep link or https)
        cancelUrl:
          type: string
          description: Return URL if the customer cancels
        metadata:
          type: object
          additionalProperties:
            type: string
    CreatePaymentLinkRequest:
      type: object
      properties:
        merchantRefPrefix:
          type: string
        payoutAddressEvm:
          type: string
        amountUsdc:
          type: number
        priceId:
          type: string
        description:
          type: string
        maxUses:
          type: integer
          nullable: true
        expiresAt:
          type: string
          format: date-time
          nullable: true
        successUrl:
          type: string
          format: uri
        cancelUrl:
          type: string
          format: uri
        collectEmail:
          type: boolean
        collectName:
          type: boolean
        metadata:
          type: object
          additionalProperties:
            type: string